Update build and release script.

+++ android/build-android.sh
@@ -7,6 +7,12 @@ function echo_y () {
     echo -e "${YELLOW}$@${RESET}"
 }
 
+# 2. check NDK
+if [ -z "$ANDROID_NDK" ]; then
+    echo_r "[android] environment variable 'ANDROID_NDK' need to be setup"
+    exit 1
+fi
+
 # remove
 rm -rf dist
 mkdir dist
@@ -17,7 +23,7 @@ cd $root/MNN
 
 # cmake . -MNN_VULKAN=ON -DMNN_BUILD_SHARED_LIBS=OFF -DMNN_PORTABLE_BUILD=ON && make -j4
 
-ARCHS="armeabi-v7a arm64-v8a x86 x86_64"
+ARCHS="armeabi-v7a arm64-v8a x86_64 " 
 
 echo_y "\nCleaning up ..."
 for ARCH in ${ARCHS}
@@ -34,27 +40,70 @@ do
         -DCMAKE_BUILD_TYPE=Release \
         -DANDROID_NATIVE_API_LEVEL=android-21  \
         -DANDROID_TOOLCHAIN=clang \
-        -DMNN_USE_LOGCAT=false \
-        -DMNN_SEP_BUILD=0 -DMNN_BUILD_SHARED_LIBS=OFF \
-        -DMNN_VULKAN=true \
+        -DMNN_USE_LOGCAT=false  \
+        -DMNN_SEP_BUILD=0  \
+        -DMNN_BUILD_SHARED_LIBS=OFF  -DMNN_VULKAN=OFF \
         -DMNN_BUILD_FOR_ANDROID_COMMAND=true \
         -DNATIVE_LIBRARY_OUTPUT=. -DNATIVE_INCLUDE_OUTPUT=. $1 $2 $3
 
     make -j4
+    mv libMNN.a $root/android/dist/libMNN-${ARCH}.a
 
+    cmake $root/MNN \
+        -DCMAKE_TOOLCHAIN_FILE=$ANDROID_NDK/build/cmake/android.toolchain.cmake \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_ABI=${ARCH} \
+        -DANDROID_STL=c++_static \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_NATIVE_API_LEVEL=android-21  \
+        -DANDROID_TOOLCHAIN=clang \
+        -DMNN_USE_LOGCAT=false  \
+        -DMNN_SEP_BUILD=0  \
+        -DMNN_BUILD_SHARED_LIBS=ON  -DMNN_VULKAN=OFF \
+        -DMNN_BUILD_FOR_ANDROID_COMMAND=true \
+        -DNATIVE_LIBRARY_OUTPUT=. -DNATIVE_INCLUDE_OUTPUT=. $1 $2 $3
+
+    make -j4
+    mv libMNN.so $root/android/dist/libMNN-${ARCH}.so
+
+    echo_y "\nBuilding android Vulkan library for ${ARCH}"
+    cmake $root/MNN \
+        -DCMAKE_TOOLCHAIN_FILE=$ANDROID_NDK/build/cmake/android.toolchain.cmake \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_ABI=${ARCH} \
+        -DANDROID_STL=c++_static \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_NATIVE_API_LEVEL=android-21  \
+        -DANDROID_TOOLCHAIN=clang \
+        -DMNN_USE_LOGCAT=false  \
+        -DMNN_SEP_BUILD=0  \
+        -DMNN_BUILD_SHARED_LIBS=OFF -DMNN_VULKAN=ON \
+        -DMNN_BUILD_FOR_ANDROID_COMMAND=true \
+        -DNATIVE_LIBRARY_OUTPUT=. -DNATIVE_INCLUDE_OUTPUT=. $1 $2 $3
+    make -j4
+    mv libMNN.a $root/android/dist/libMNN_Vulkan-${ARCH}.a
+
+    echo_y "\nBuilding android Vulkan shared library for ${ARCH}"
+    cmake $root/MNN \
+        -DCMAKE_TOOLCHAIN_FILE=$ANDROID_NDK/build/cmake/android.toolchain.cmake \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_ABI=${ARCH} \
+        -DANDROID_STL=c++_static \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DANDROID_NATIVE_API_LEVEL=android-21  \
+        -DANDROID_TOOLCHAIN=clang \
+        -DMNN_USE_LOGCAT=false  \
+        -DMNN_SEP_BUILD=0  \
+        -DMNN_BUILD_SHARED_LIBS=ON -DMNN_VULKAN=ON \
+        -DMNN_BUILD_FOR_ANDROID_COMMAND=true \
+        -DNATIVE_LIBRARY_OUTPUT=. -DNATIVE_INCLUDE_OUTPUT=. $1 $2 $3
+    make -j4
+    mv libMNN.so $root/android/dist/libMNN_Vulkan-${ARCH}.so
+
+    cd $root/android
+    rm -rf build_$ARCH
 done
 #rm -rf build_32 build_64
-
-#echo_y "\nBuilding android shared library for 32bit archs"
-#mkdir build_32 && cd build_32
-
-
-echo_y "\nBuilding android shared library for 64bit archs"
-#cd $root/MNN/project/android
-#mkdir build_64 && cd build_64 && ../build_64.sh
-
-echo_y "\nCopying files"
-
 
 echo_y "\nAndroid build finished"
